Improve border look for selected region.
Removes `bringToFront` call on mouseovers, so that the border for the selected
region remains intact. Otherwise, when mousing away after selecting a region,
the `bringTofront` on an adjacent region would cause the selected region border
to get thinner.

Not calling `bringToFront` on mouseover means the border highlight can look a
bit wonky, but it's better to have the selected region border always look good.